groups activities to activity.js
