+++ routes/api/group.js
@@ -1132,308 +1132,7 @@ module.exports = function (app) {
             .catch(next);
     });
 
-    /**
-     * Read (List) public Group Activities
-     */
-
-    var groupActivitiesList = function (req, res, next, visibility) {
-        var limitMax = 50;
-        var limitDefault = 10;
-        var groupId = req.params.groupId;
-        var userId = null;
-        if (req.user) {
-            userId = req.user.id;
-        }
-        var offset = parseInt(req.query.offset, 10) ? parseInt(req.query.offset, 10) : 0;
-        var limit = parseInt(req.query.limit, 10) ? parseInt(req.query.limit, 10) : limitDefault;
-        var visibilityCondition = '';
-        if (visibility) {
-            visibilityCondition = 'g.visibility = :visibility AND';
-        }
-
-        if (limit > limitMax) limit = limitDefault;
-
-        return db.transaction(function (t) {
-            var activity = Activity.build({
-                data: {
-                    offset: offset,
-                    limit: limit
-                }
-            });
-
-            return db
-                .query('\
-                    CREATE OR REPLACE FUNCTION pg_temp.parseData(data jsonb) \
-                        RETURNS jsonb \
-                        AS \
-                        $BODY$ \
-                        DECLARE \
-                        finalData jsonb = data; \
-                        BEGIN \
-                            IF ((data ? \'actor\') AND data#>>\'{actor, type}\' = \'User\' AND data#>>\'{actor, id}\' IS NOT NULL AND (data ? \'object\' AND data#>>\'{object, 0, @type}\' = \'VoteList\')) THEN \
-                                SELECT jsonb_set( \
-                                    data, \
-                                    \'{actor}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'type\', data#>>\'{actor, type}\', \
-                                            \'name\', \'User\', \
-                                            \'company\', \'\' \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Users" u WHERE u.id::text = data#>>\'{actor, id}\'; \
-                            ELSIF ((data ? \'actor\') AND data#>>\'{actor, type}\' = \'User\' AND data#>>\'{actor, id}\' IS NOT NULL ) THEN \
-                                SELECT jsonb_set( \
-                                    data, \
-                                    \'{actor}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', u.id, \
-                                            \'type\', data#>>\'{actor, type}\', \
-                                            \'name\', u.name, \
-                                            \'company\', u.company \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Users" u WHERE u.id::text = data#>>\'{actor, id}\'; \
-                            ELSE \
-                                SELECT jsonb_set( \
-                                    data, \
-                                    \'{actor}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'type\', data#>>\'{actor, type}\' \
-                                        ) \
-                                    ), \
-                                false) INTO finalData; \
-                            END IF; \
-                            IF ((data ? \'actor\') AND data#>>\'{actor, type}\' = \'User\' AND data#>>\'{actor, id}\' IS NULL  AND (data ? \'object\' AND data#>>\'{object, 0, @type}\' = \'VoteList\')) THEN \
-                                SELECT \
-                                    jsonb_set( \
-                                        data, \
-                                        \'{object, 0}\', \
-                                        to_jsonb( \
-                                            (data#>>\'{object, 0}\')::jsonb - \'userId\' \
-                                        ), \
-                                    false ) INTO finalData; \
-                            END IF; \
-                            IF ((finalData ? \'object\') AND finalData#>>\'{object, @type}\' = \'Group\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{object}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', g.id, \
-                                            \'@type\', finalData#>>\'{object, @type}\', \
-                                            \'parentId\', g."parentId", \
-                                            \'name\', g.name, \
-                                            \'creatorId\', g."creatorId", \
-                                            \'visibility\', g.visibility, \
-                                            \'sourcePartnerId\', g."sourcePartnerId", \
-                                            \'createdAt\', g."createdAt", \
-                                            \'updatedAt\', g."updatedAt", \
-                                            \'deletedAt\', g."deletedAt" \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Groups" g WHERE g.id::text = data#>>\'{object, id}\'; \
-                            END IF; \
-                            IF ((finalData ? \'object\') AND finalData#>>\'{object, @type}\' = \'Topic\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{object}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', t.id, \
-                                            \'@type\', finalData#>>\'{object, @type}\', \
-                                            \'title\', t.title, \
-                                            \'status\', t.status, \
-                                            \'visibility\', t.visibility, \
-                                            \'categories\', t.categories, \
-                                            \'sourcePartnerId\', t."sourcePartnerId", \
-                                            \'creatorId\', t."creatorId", \
-                                            \'tokenJoin\', t."tokenJoin", \
-                                            \'padUrl\', t."padUrl", \
-                                            \'endsAt\', t."endsAt", \
-                                            \'hashtag\', t.hashtag, \
-                                            \'createdAt\', t."createdAt", \
-                                            \'updatedAt\', t."updatedAt" \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Topics" t WHERE t.id::text = data#>>\'{object, id}\'; \
-                            END IF; \
-                            IF ((finalData ? \'object\') AND finalData#>>\'{object, @type}\' = \'User\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{object}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', u.id, \
-                                            \'@type\', finalData#>>\'{object, @type}\', \
-                                            \'name\', u.name, \
-                                            \'company\', u."company" \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Users" u WHERE u.id::text = data#>>\'{object, id}\'; \
-                            END IF; \
-                            IF ((finalData ? \'origin\') AND finalData#>>\'{origin, @type}\' = \'TopicMemberUser\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{object}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'@type\', finalData#>>\'{origin, @type}\', \
-                                            \'level\', finalData#>>\'{origin, level}\', \
-                                            \'userId\', finalData#>>\'{origin, userId}\', \
-                                            \'topicId\', finalData#>>\'{origin, topicId}\', \
-                                            \'createdAt\', finalData#>>\'{origin, createdAt}\', \
-                                            \'deletedAt\', finalData#>>\'{origin, deletedAt}\', \
-                                            \'updatedAt\', finalData#>>\'{origin, updatedAt}\', \
-                                            \'userName\', u.name, \
-                                            \'topicTitle\', t.title \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Users" u JOIN "Topics" t ON t.id = t.id WHERE u.id::text = data#>>\'{origin, userId}\' AND t.id::text = data#>>\'{origin, topicId}\'; \
-                            END IF; \
-                            IF ((finalData ? \'origin\') AND finalData#>>\'{origin, @type}\' = \'TopicMemberGroup\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{object}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'@type\', finalData#>>\'{origin, @type}\', \
-                                            \'level\', finalData#>>\'{origin, level}\', \
-                                            \'userId\', finalData#>>\'{origin, userId}\', \
-                                            \'topicId\', finalData#>>\'{origin, topicId}\', \
-                                            \'createdAt\', finalData#>>\'{origin, createdAt}\', \
-                                            \'deletedAt\', finalData#>>\'{origin, deletedAt}\', \
-                                            \'updatedAt\', finalData#>>\'{origin, updatedAt}\', \
-                                            \'groupName\', g.name, \
-                                            \'topicTitle\', t.title \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Groups" g JOIN "Topics" t ON t.id = t.id WHERE g.id::text = data#>>\'{origin, userId}\' AND t.id::text = data#>>\'{origin, topicId}\'; \
-                            END IF; \
-                            IF ((finalData ? \'target\') AND finalData#>>\'{target, @type}\' = \'Group\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{target}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', g.id, \
-                                            \'@type\', finalData#>>\'{target, @type}\', \
-                                            \'parentId\', g."parentId", \
-                                            \'name\', g.name, \
-                                            \'creatorId\', g."creatorId", \
-                                            \'visibility\', g.visibility, \
-                                            \'sourcePartnerId\', g."sourcePartnerId", \
-                                            \'createdAt\', g."createdAt", \
-                                            \'updatedAt\', g."updatedAt", \
-                                            \'deletedAt\', g."deletedAt" \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Groups" g WHERE g.id::text = data#>>\'{target, id}\'; \
-                            END IF; \
-                            IF ((finalData ? \'target\') AND finalData#>>\'{target, @type}\' = \'Topic\') THEN \
-                                SELECT jsonb_set( \
-                                    finalData, \
-                                    \'{target}\', \
-                                    to_jsonb( \
-                                        json_build_object( \
-                                            \'id\', t.id, \
-                                            \'@type\', finalData#>>\'{target, @type}\', \
-                                            \'title\', t.title, \
-                                            \'status\', t.status, \
-                                            \'visibility\', t.visibility, \
-                                            \'categories\', t.categories, \
-                                            \'sourcePartnerId\', t."sourcePartnerId", \
-                                            \'creatorId\', t."creatorId", \
-                                            \'tokenJoin\', t."tokenJoin", \
-                                            \'padUrl\', t."padUrl", \
-                                            \'endsAt\', t."endsAt", \
-                                            \'hashtag\', t.hashtag, \
-                                            \'createdAt\', t."createdAt", \
-                                            \'updatedAt\', t."updatedAt" \
-                                        ) \
-                                    ), \
-                                false) INTO finalData FROM "Topics" t WHERE t.id::text = data#>>\'{target, id}\'; \
-                            END IF; \
-                            RETURN finalData; \
-                        END; \
-                        $BODY$ \
-                        LANGUAGE plpgsql; \
-                    \
-                    SELECT \
-                        a.id, \
-                        pdata as data, \
-                        a."createdAt", \
-                        a."updatedAt", \
-                        a."deletedAt" \
-                    FROM \
-                        "Activities" a, \
-                        pg_temp.parseData(a.data) pdata \
-                        JOIN "Groups" g ON g.id = :groupId \
-                        WHERE \
-                        ' + visibilityCondition + ' \
-                        ARRAY[:groupId] <@  a."groupIds" \
-                        OR \
-                        a.data@>\'{"type": "View"}\' \
-                        AND \
-                        a."actorType" = \'User\' \
-                        AND \
-                        a."actorId" = :userId \
-                        GROUP BY a.id, pdata, a."createdAt", a."updatedAt", a."deletedAt" \
-                        ORDER BY a."updatedAt" DESC \
-                        LIMIT :limit OFFSET :offset \
-            ;', {
-                    replacements: {
-                        groupId: groupId,
-                        userId: userId,
-                        visibility: visibility,
-                        limit: limit,
-                        offset: offset
-                    },
-                    type: db.QueryTypes.SELECT,
-                    transaction: t,
-                    raw: true
-                })
-                .then(function (results) {
-                    if (userId) {
-                        return cosActivities
-                            .viewActivityFeedActivity(
-                                activity,
-                                {
-                                    type: 'User',
-                                    id: req.user.id
-                                },
-                                req.method + ' ' + req.path,
-                                t
-                            )
-                            .then(function () {
-                                return results;
-                            });
-                    }
-
-                    return results;
-                });
-        });
+    return {
+        hasPermission: hasPermission
     };
-
-    app.get('/api/groups/:groupId/activities', function (req, res, next) {
-        return groupActivitiesList(req, res, next, 'public')
-            .then(function (results) {
-                if (results && results.length && results[0]) {
-                    return res.ok(results);
-                } else {
-                    return res.notFound();
-                }
-            })
-            .catch(next);
-    });
-
-    app.get('/api/users/:userId/groups/:groupId/activities', loginCheck(['partner']), hasPermission(GroupMember.LEVELS.read, true), function (req, res, next) {
-        return groupActivitiesList(req, res, next)
-            .then(function (results) {
-                return res.ok(results);
-            })
-            .catch(next);
-    });
 };
